requirements.txt
numpy==1.17.2
pytest==5.2.1
matplotlib==3.1.1
PyYAML==5.3
